feat: 문화력 조회, 상세정보 확인, 삭제 기능 완료 (영화)
자신이 등록한 문화력을 조회하고 확인할 수 있음. 삭제하면 db까지 삭제.

+++ backend/src/main/java/com/ssafy/api/request/CulturePostReq.java
@@ -17,7 +17,7 @@ import java.util.List;
 public class CulturePostReq {
 
     @ApiModelProperty(name = "책(1) 또는 영화(0)", example = "0")
-    int cultureCategory;
+    Long cultureCategory;
 
     @ApiModelProperty(name = "책/영화 제목", example = "7년의 밤")
     String cultureTitle;
