+++ backend/src/main/java/com/ssafy/api/service/CultureService.java
@@ -8,5 +8,5 @@ import java.util.List;
 public interface CultureService {
     List<CultureListRes> getCultureInfo(String email);
     void postCulture(CulturePostReq culturePostReq, String email);
-    void deleteCulture(String email, String cultureTitle, int cultureCategory);
+    void deleteCulture(String email, String cultureTitle, Long cultureCategory);
 }
